backend/requirements.txt
fastapi>=0.109.0
python-multipart>=0.0.7
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
cryptography>=41.0.7
websockets>=12.0
python-pptx>=1.0,<2.0
pypdf>=3.17,<4.0
weasyprint>=65,<70
Pillow>=10.0.0
alembic>=1.13.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
numpy>=1.26.0
scipy>=1.12.0
PyYAML>=6.0
